General tidying. Remove superfluous logical checks. Introduce switch.
The multiline logic for the missingHitType declaration is moved into named constants
to improve readability/maintainability. The if/else block checking
whether the missingHitType is TPC_VIEW_W has been moved to a
switch-case. The block also handles the previous if-statement that
throws a STATUS_CODE_INVALID_PARAMETERS exception if the missingHitType
is HIT_CUSTOM.